Move print-file-metadata over to blockstore subcommand
