+++ ledger-tool/src/main.rs
@@ -37,7 +37,6 @@ use {
     solana_ledger::{
         ancestor_iterator::AncestorIterator,
         blockstore::{create_new_ledger, Blockstore, PurgeType},
-        blockstore_db::{columns as cf, Column, ColumnName},
         blockstore_options::{AccessType, LedgerColumnOptions, BLOCKSTORE_DIRECTORY_ROCKS_FIFO},
         blockstore_processor::ProcessOptions,
         shred::Shred,
@@ -481,74 +480,6 @@ fn graph_forks(bank_forks: &BankForks, config: &GraphConfig) -> String {
 
     dot.push("}".to_string());
     dot.join("\n")
-}
-
-fn raw_key_to_slot(key: &[u8], column_name: &str) -> Option<Slot> {
-    match column_name {
-        cf::SlotMeta::NAME => Some(cf::SlotMeta::slot(cf::SlotMeta::index(key))),
-        cf::Orphans::NAME => Some(cf::Orphans::slot(cf::Orphans::index(key))),
-        cf::DeadSlots::NAME => Some(cf::SlotMeta::slot(cf::SlotMeta::index(key))),
-        cf::DuplicateSlots::NAME => Some(cf::SlotMeta::slot(cf::SlotMeta::index(key))),
-        cf::ErasureMeta::NAME => Some(cf::ErasureMeta::slot(cf::ErasureMeta::index(key))),
-        cf::BankHash::NAME => Some(cf::BankHash::slot(cf::BankHash::index(key))),
-        cf::Root::NAME => Some(cf::Root::slot(cf::Root::index(key))),
-        cf::Index::NAME => Some(cf::Index::slot(cf::Index::index(key))),
-        cf::ShredData::NAME => Some(cf::ShredData::slot(cf::ShredData::index(key))),
-        cf::ShredCode::NAME => Some(cf::ShredCode::slot(cf::ShredCode::index(key))),
-        cf::TransactionStatus::NAME => Some(cf::TransactionStatus::slot(
-            cf::TransactionStatus::index(key),
-        )),
-        cf::AddressSignatures::NAME => Some(cf::AddressSignatures::slot(
-            cf::AddressSignatures::index(key),
-        )),
-        cf::TransactionMemos::NAME => None, // does not implement slot()
-        cf::TransactionStatusIndex::NAME => None, // does not implement slot()
-        cf::Rewards::NAME => Some(cf::Rewards::slot(cf::Rewards::index(key))),
-        cf::Blocktime::NAME => Some(cf::Blocktime::slot(cf::Blocktime::index(key))),
-        cf::PerfSamples::NAME => Some(cf::PerfSamples::slot(cf::PerfSamples::index(key))),
-        cf::BlockHeight::NAME => Some(cf::BlockHeight::slot(cf::BlockHeight::index(key))),
-        cf::ProgramCosts::NAME => None, // does not implement slot()
-        cf::OptimisticSlots::NAME => {
-            Some(cf::OptimisticSlots::slot(cf::OptimisticSlots::index(key)))
-        }
-        &_ => None,
-    }
-}
-
-fn print_blockstore_file_metadata(
-    blockstore: &Blockstore,
-    file_name: &Option<&str>,
-) -> Result<(), String> {
-    let live_files = blockstore
-        .live_files_metadata()
-        .map_err(|err| format!("{err:?}"))?;
-
-    // All files under live_files_metadata are prefixed with "/".
-    let sst_file_name = file_name.as_ref().map(|name| format!("/{name}"));
-    for file in live_files {
-        if sst_file_name.is_none() || file.name.eq(sst_file_name.as_ref().unwrap()) {
-            println!(
-                "[{}] cf_name: {}, level: {}, start_slot: {:?}, end_slot: {:?}, size: {}, \
-                 num_entries: {}",
-                file.name,
-                file.column_family_name,
-                file.level,
-                raw_key_to_slot(&file.start_key.unwrap(), &file.column_family_name),
-                raw_key_to_slot(&file.end_key.unwrap(), &file.column_family_name),
-                file.size,
-                file.num_entries,
-            );
-            if sst_file_name.is_some() {
-                return Ok(());
-            }
-        }
-    }
-    if sst_file_name.is_some() {
-        return Err(format!(
-            "Failed to find or load the metadata of the specified file {file_name:?}"
-        ));
-    }
-    Ok(())
 }
 
 fn compute_slot_cost(blockstore: &Blockstore, slot: Slot) -> Result<(), String> {
@@ -1837,23 +1768,6 @@ fn main() {
                         ),
                 ),
         )
-        .subcommand(
-            SubCommand::with_name("print-file-metadata")
-                .about(
-                    "Print the metadata of the specified ledger-store file. If no file name is \
-                     specified, it will print the metadata of all ledger files.",
-                )
-                .arg(
-                    Arg::with_name("file_name")
-                        .long("file-name")
-                        .takes_value(true)
-                        .value_name("SST_FILE_NAME")
-                        .help(
-                            "The ledger file name (e.g. 011080.sst.) If no file name is \
-                             specified, it will print the metadata of all ledger files.",
-                        ),
-                ),
-        )
         .program_subcommand()
         .get_matches();
 
@@ -1880,6 +1794,7 @@ fn main() {
         | ("bounds", Some(_))
         | ("dead-slots", Some(_))
         | ("duplicate-slots", Some(_))
+        | ("print-file-metadata", Some(_))
         | ("remove-dead-slot", Some(_))
         | ("set-dead-slot", Some(_)) => blockstore_process_command(&ledger_path, &matches),
         _ => {
@@ -3656,14 +3571,6 @@ fn main() {
                         if let Err(err) = compute_slot_cost(&blockstore, slot) {
                             eprintln!("{err}");
                         }
-                    }
-                }
-                ("print-file-metadata", Some(arg_matches)) => {
-                    let blockstore =
-                        open_blockstore(&ledger_path, arg_matches, AccessType::Secondary);
-                    let sst_file_name = arg_matches.value_of("file_name");
-                    if let Err(err) = print_blockstore_file_metadata(&blockstore, &sst_file_name) {
-                        eprintln!("{err}");
                     }
                 }
                 ("", _) => {
